fix: web pck export no longer borrows the import target's inputs
Review finding: TwoDogExportPack used @(TwoDogImportInputs) for
incrementality, but that item is only populated when TwoDogAutoImport is
enabled - with the documented TwoDogAutoImport=false knob, MSBuild skipped
the incremental target (empty Inputs) and the bundle silently shipped
without godot.pck. The export now has its own input glob over the game
project sources. Verified: clean publish with -p:TwoDogAutoImport=false
produces and deploys the pck.

Also renames the target to TwoDogExportGamePack (it shared its name with
the TwoDogExportPack boolean property), and adds ExportPackToolTests
covering the helper's --export-pack mode in-process (pck produced, GDPC
magic, missing --output fails), with the import-tool test plumbing
extracted into a shared HelperToolTestBed.
